refactor: apply DTO projection to instagram search APIs for central, union, college, and department
